* installed expo ImageManipulator to compress photos before sending for faster upload and download and less data consumption  * created custom cached image component that fixed android memory issue causing app to become not responsive when list of images is large * Implemented screen shotting of camera to increase speed of picture being taken and fix jittering of pic after being taken * implemented react-native-big-list, a faster list library that recycles views instead of creating new ones so memory consumption is drastically reduced  * Created and implemented a scroll component for when picture is opened inside a gallery, so pictures within the gallery can we continuously scrolled in full size. * worked with bank end to fix bug that returns thumbnails in landscape mode.

+++ src/components/ActionBottomSheet.js
@@ -73,19 +73,17 @@ const ActionBottomSheet = forwardRef((props, ref) => {
 
     //----------------------------------------------------------------DELETING GALLERY----------------------------------------------------------------
     const deleteGalleryHandler = useCallback(async () => {
-        const wholeDirectory =
-            FileSystem.cacheDirectory + 'images/' + `${props.galleryID}t`
+        // const wholeDirectory = FileSystem.cacheDirectory + 'images/'
         const picDir =
-            FileSystem.cacheDirectory + 'images/' + `${props.galleryID}t`
-        const removePic = await FileSystem.deleteAsync(picDir)
-        console.log(
-            '🚀 ~ file: ActionBottomSheet.js ~ line 78 ~ deleteGalleryHandler ~ removePic',
-            removePic
-        )
+            FileSystem.cacheDirectory + 'images/' + `${props.galleryID.id}t`
+        const cacheExists = await FileSystem.getInfoAsync(picDir)
+        if (cacheExists.exists) {
+            await FileSystem.deleteAsync(picDir)
+        }
 
         await dispatch(deleteGallery(props.galleryID))
-        await await FileSystem.deleteAsync(picDir)
         bottomSheetModalRef.current?.close()
+        props.refreshGalleryList()
     }, [props])
     //----------------------------------------------------------------DELETING GALLERY----------------------------------------------------------------
